requirements.txt
python-dotenv==1.0.1
discord.py==2.3.2